requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
openai==1.12.0
